% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_decision.R
\name{add_decision}
\alias{add_decision}
\title{add decision}
\usage{
add_decision(decision_obj, course_id, year, term, decision, notes)
}
\arguments{
\item{decision_obj}{decision in object}

\item{course_id}{course ID}

\item{year}{offering year}

\item{term}{offering term}

\item{decision}{final decision}

\item{notes}{any notes}
}
\value{
decision object
}
\description{
add decision
}
