% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_example.R
\name{add_example}
\alias{add_example}
\title{add example}
\usage{
add_example(folder, type)
}
\arguments{
\item{folder}{folder to place example}

\item{type}{type of example file}
}
\description{
add example
}
