% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_names.R
\name{add_names}
\alias{add_names}
\title{add names}
\usage{
add_names(gr, course_names)
}
\arguments{
\item{gr}{graderosters}

\item{course_names}{tibble with course id and course names from peoplesoft}
}
\value{
graderosters with added names
}
\description{
add names
}
