% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assessment_obj}
\alias{assessment_obj}
\title{assessment_obj}
\format{
A data frame with 1,600 rows and 11 variables
\describe{
\item{id}{Student ID}
\item{name}{Student Name}
\item{course_id}{Course ID of form DISPLINE-CAT_NO}
\item{year}{Year of offering}
\item{term}{Term of offering}
\item{mark}{Mark for course NA for WNF etc.}
\item{grade}{Grade}
\item{assessment}{Assessment piece}
\item{value}{Assessment mark}
\item{total}{Assessment total marks}
\item{p}{Proportion}
}
}
\source{
"Simulated data"
}
\usage{
assessment_obj
}
\description{
A simulated tibble showing the standard assessment-object form
}
\keyword{datasets}
