% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_mark_obj.R
\name{augment_mark_obj}
\alias{augment_mark_obj}
\title{augment_mark_obj}
\usage{
augment_mark_obj(mark_obj)
}
\arguments{
\item{mark_obj}{mark-obj}
}
\value{
augmented-mark-obj
}
\description{
augment_mark_obj
}
\examples{
augment_mark_obj(example_marks)
}
