% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented_example_marks.R
\docType{data}
\name{augmented_example_marks}
\alias{augmented_example_marks}
\title{augmented_example_marks}
\format{
A data frame with 24,431 rows and 11 variables
\describe{
\item{id}{Student ID}
\item{name}{Student Name}
\item{course_name}{Course Name}
\item{course_id}{Course ID of form DISPLINE-CAT_NO}
\item{year}{Year of offering}
\item{term}{Term of offering}
\item{mark}{Mark for course NA for WNF etc.}
\item{grade}{Grade}
\item{raa}{RAA note}
\item{N}{Number of exams with a mark for a student}
\item{mean}{Mean exam mark}
\item{diff}{Difference of mark from mean of all over marks for a given student}
}
}
\source{
"Simulated data"
}
\usage{
augmented_example_marks
}
\description{
A simulated tibble showing the standard mark-object form
}
\keyword{datasets}
