% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_grade.R
\name{calc_grade}
\alias{calc_grade}
\title{calc_grade}
\usage{
calc_grade(x)
}
\arguments{
\item{x}{vector of marks.}
}
\value{
vector of grades
}
\description{
Based on https://www.adelaide.edu.au/policies/700/all/?dsn=policy.version;field=data;id=34226;m=view
}
\examples{
calc_grade(c("FNS", "WNF", "RP", "WF", 0:100))
}
