% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-gradebook.R
\name{clean_gradebook}
\alias{clean_gradebook}
\title{clean gradebook}
\usage{
clean_gradebook(file)
}
\arguments{
\item{file}{gradebook CSV}
}
\value{
cleaned gradebook
}
\description{
clean gradebook
}
