% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{clean names}
\usage{
clean_names(x)
}
\arguments{
\item{x}{vector of column names}
}
\value{
clean vector of names
}
\description{
code to clean names of gradebook
}
\examples{
clean_names("Assignment (123)")
}
