% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_term.R
\name{convert_term}
\alias{convert_term}
\title{convert term}
\usage{
convert_term(x)
}
\arguments{
\item{x}{vector of term codes}
}
\value{
vector of term names
}
\description{
convert the term codes to term names
}
\examples{
convert_term(4320)
}
