% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_acad_supps.R
\name{count_acad_supps}
\alias{count_acad_supps}
\title{count_acad_supps}
\usage{
count_acad_supps(mark_obj, course_id, year, term)
}
\arguments{
\item{mark_obj}{mark obj}

\item{course_id}{course id}

\item{year}{year}

\item{term}{term}
}
\value{
number of students with mark between 45 and 49
}
\description{
count_acad_supps
}
