% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_decision_df.R
\name{create_decision_obj}
\alias{create_decision_obj}
\title{create decision obj}
\usage{
create_decision_obj(file, term)
}
\arguments{
\item{file}{file of decision CSV}

\item{term}{term to start with}
}
\value{
decision_obj
}
\description{
create decision obj
}
