% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_running_list.R
\name{create_running_list}
\alias{create_running_list}
\title{create running list}
\usage{
create_running_list(mark_obj, year, term)
}
\arguments{
\item{mark_obj}{mark object}

\item{year}{year of examiners meeting}

\item{term}{term of examiners meeting}
}
\value{
list of running list
}
\description{
create running list
}
