% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_ex.R
\name{deal_ex}
\alias{deal_ex}
\title{deal with EX}
\usage{
deal_ex(x)
}
\arguments{
\item{x}{vector of marks}
}
\value{
cleaned vector of marks
}
\description{
converts missing to 0, and EX to missing
}
\examples{
deal_ex(c(1, NA, "EX"))
}
