% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decision_obj}
\alias{decision_obj}
\title{decision obj}
\format{
A data frame with 21 rows and 6 variables:
\describe{
\item{time}{Time of decision}
\item{course_id}{Course ID}
\item{year}{Year of offering}
\item{term}{Term of offering}
\item{decision}{Final decision}
\item{notes}{Notes}
}
}
\source{
"Simulated"
}
\usage{
decision_obj
}
\description{
A dataset containing decisions from examiners' meeting
}
\keyword{datasets}
