% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examiners_reports}
\alias{examiners_reports}
\title{examiners_reports}
\format{
A data frame with 1 row and 13 variables
\describe{
\item{course_cc}{Course coordinator}
\item{course_name}{Course Name}
\item{assessment}{Assessment structure}
\item{audit}{Audit done}
\item{hurdle}{Course hurdle}
\item{hurdle_details}{Hurdle details}
\item{stack}{Co-taught courses}
\item{scaling}{Course scaling}
\item{scaling_details}{Course scaling details}
\item{student}{Student to discuss}
\item{comments}{Comments}
\item{changes}{Proposed changes}
\item{course_id}{Course ID of form DISPLINE-CAT_NO}
}
}
\source{
"Simulated data"
}
\usage{
examiners_reports
}
\description{
A simulated tibble showing the standard mark-object form
}
\keyword{datasets}
