% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_marks.R
\docType{data}
\name{example_marks}
\alias{example_marks}
\title{example_marks}
\format{
A data frame with 24,431 rows and 8 variables
\describe{
\item{id}{Student ID}
\item{name}{Student Name}
\item{course_name}{Course Name}
\item{course_id}{Course ID of form DISPLINE-CAT_NO}
\item{year}{Year of offering}
\item{term}{Term of offering}
\item{mark}{Mark for course NA for WNF etc.}
\item{grade}{Grade}
\item{raa}{RAA note}
}
}
\source{
"Simulated data"
}
\usage{
example_marks
}
\description{
A simulated tibble showing the standard mark-object form
}
\keyword{datasets}
