% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_graderoster.R
\name{fill_graderoster}
\alias{fill_graderoster}
\title{fill grade roster}
\usage{
fill_graderoster(mark_obj, infile, year, term, RAA = 45:49)
}
\arguments{
\item{mark_obj}{mark object}

\item{infile}{empty graderoster}

\item{year}{offering year}

\item{term}{offering term}

\item{RAA}{marks to give RAA}
}
\value{
nothing
}
\description{
fill grade roster
}
