% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_assessment.R
\name{filter_assessment}
\alias{filter_assessment}
\title{filter assessment}
\usage{
filter_assessment(assessment_obj, course_id, year, term)
}
\arguments{
\item{assessment_obj}{assessment object}

\item{course_id}{course_id}

\item{year}{year of offering}

\item{term}{year of offering}
}
\value{
filtered assessment object
}
\description{
filter assessment
}
\examples{
filter_assessment(assessment_obj, "STATS-3022", 2023, "Sem 2")
}
