% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assessment_list.R
\name{get_assessment_list}
\alias{get_assessment_list}
\title{get assessment list}
\usage{
get_assessment_list(assessment_obj)
}
\arguments{
\item{assessment_obj}{assessment object}
}
\value{
list of assessments
}
\description{
get assessment list
}
\examples{
get_assessment_list(assessment_obj)
}
