% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decision_dt.R
\name{get_decision_dt}
\alias{get_decision_dt}
\title{get decision datatable}
\usage{
get_decision_dt(decision_obj, year, term)
}
\arguments{
\item{decision_obj}{decision object}

\item{year}{year to show}

\item{term}{term to show}
}
\value{
datatable
}
\description{
get decision datatable
}
