% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_examiners_report.R
\name{get_examiners_report}
\alias{get_examiners_report}
\title{get examiners report}
\usage{
get_examiners_report(examiners_reports, examiner)
}
\arguments{
\item{examiners_reports}{examiners report tibble}

\item{examiner}{course coordinator}
}
\value{
table of report
}
\description{
get examiners report
}
