% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grade_tab.R
\name{get_grade_tab}
\alias{get_grade_tab}
\title{get_grade_tab}
\usage{
get_grade_tab(mark_obj, course_id, term)
}
\arguments{
\item{mark_obj}{standard mark tibble}

\item{course_id}{course-id}

\item{term}{term}
}
\value{
gt-table
}
\description{
Gives a table of grades
}
\examples{
get_grade_tab(mark_obj, course_id = "STATS-3001", term = "Sem 1")
}
