% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graderoster_course_id.R
\name{get_graderoster_course_id}
\alias{get_graderoster_course_id}
\title{get GR course id}
\usage{
get_graderoster_course_id(file)
}
\arguments{
\item{file}{graderoster file}
}
\value{
course id
}
\description{
get GR course id
}
