% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{get_ids}
\alias{get_ids}
\title{get ids}
\usage{
get_ids(augmented_mark_obj, course_id, year, term, rows)
}
\arguments{
\item{augmented_mark_obj}{augmented mark obj}

\item{course_id}{course id}

\item{year}{year of offering}

\item{term}{term of offering}

\item{rows}{selected rows}
}
\value{
IDs selected
}
\description{
get ids
}
