% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intersection.R
\name{get_intersection}
\alias{get_intersection}
\title{get_intersection}
\usage{
get_intersection(mark_obj, course_id, term, year, other_course_ids)
}
\arguments{
\item{mark_obj}{mark-obj}

\item{course_id}{course-id}

\item{term}{term}

\item{year}{year of offering}

\item{other_course_ids}{course to compare with}
}
\value{
tibble of marks for students in first course
}
\description{
Gives a table of marks
}
