% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intersection_dt.R
\name{get_intersection_dt}
\alias{get_intersection_dt}
\title{get_intersection DT}
\usage{
get_intersection_dt(intersection_obj)
}
\arguments{
\item{intersection_obj}{intersection object}
}
\value{
tibble of marks for students in first course
}
\description{
Gives a table of marks
}
