% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mark.R
\name{get_mark}
\alias{get_mark}
\title{get mark}
\usage{
get_mark(x)
}
\arguments{
\item{x}{vector of marks}
}
\value{
numeric vector of marks
}
\description{
get mark
}
\examples{
get_mark(c(0:100, "CN", "F", "FNS", "NFE", "RP", "WF", "WNF", "NGP"))
}
