% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mark_summary.R
\name{get_mark_summary_gt}
\alias{get_mark_summary_gt}
\title{get_mark_summary}
\usage{
get_mark_summary_gt(augmented_mark_obj, course_id, term, year)
}
\arguments{
\item{augmented_mark_obj}{augmented-mark-obj}

\item{course_id}{course-id}

\item{term}{term}

\item{year}{year of offering}
}
\value{
gt summary table
}
\description{
Gives a table of marks
}
