% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mark_tab.R
\name{get_mark_tab}
\alias{get_mark_tab}
\title{get_mark_tab}
\usage{
get_mark_tab(augmented_mark_obj, course_id, term, year, hurdle = FALSE)
}
\arguments{
\item{augmented_mark_obj}{augmented-mark-obj}

\item{course_id}{course-id}

\item{term}{term}

\item{year}{year of offering}

\item{hurdle}{show only 49s}
}
\value{
DT-obj
}
\description{
Gives a table of marks
}
\examples{
get_mark_tab(augmented_mark_obj, course_id = "DISCP-3001", term = "Sem 1", 2022)
}
