% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_student_dt.R
\name{get_student_dt}
\alias{get_student_dt}
\title{get student dt}
\usage{
get_student_dt(mark_obj, ids)
}
\arguments{
\item{mark_obj}{mark object}

\item{ids}{ids}
}
\value{
gt table of student marks
}
\description{
get student dt
}
\examples{
get_student_dt(mark_obj, c("2082983", "4547752"))
}
