% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_student_tab.R
\name{get_student_tab}
\alias{get_student_tab}
\title{get student tab}
\usage{
get_student_tab(mark_obj, ids)
}
\arguments{
\item{mark_obj}{mark object}

\item{ids}{ids}
}
\value{
gt table of student marks
}
\description{
get student tab
}
