% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outlier.R
\name{is_outlier}
\alias{is_outlier}
\title{is outlier}
\usage{
is_outlier(x)
}
\arguments{
\item{x}{vector of proportions}
}
\value{
boolean if any of values are outliers
}
\description{
is outlier
}
\examples{
is_outlier(c(runif(10, 0.9, 1), 0, NA))
}
