% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_peoplesoft_valid.R
\name{is_peoplesoft_valid}
\alias{is_peoplesoft_valid}
\title{is_peoplesoft_valid}
\usage{
is_peoplesoft_valid(file)
}
\arguments{
\item{file}{peoplesoft-file}
}
\value{
boolean
}
\description{
Checks that XLSX has correct column names
}
