% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mark_obj}
\alias{mark_obj}
\title{mark_obj}
\format{
A data frame with 24,431 rows and 8 variables
\describe{
\item{id}{Student ID}
\item{name}{Student Name}
\item{course_name}{Course Name}
\item{course_id}{Course ID of form DISPLINE-CAT_NO}
\item{year}{Year of offering}
\item{term}{Term of offering}
\item{mark}{Mark for course NA for WNF etc.}
\item{grade}{Grade}
\item{raa}{RAA note}
\item{source}{where the data comes from: peoplesoft, GR, etc}
}
}
\source{
"Simulated data"
}
\usage{
mark_obj
}
\description{
A simulated tibble showing the standard mark-object form
}
\keyword{datasets}
