% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marks_dist}
\alias{marks_dist}
\title{mark_dist}
\format{
A data frame with 101 rows and 2 variables:
\describe{
\item{mark}{Number of students}
\item{n}{Number of observations}
}
}
\source{
"University of Adelaide"
}
\usage{
marks_dist
}
\description{
A dataset containing empirical mark distribution based on
UoA results
}
\keyword{datasets}
