% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_example.R
\name{move_example}
\alias{move_example}
\title{move example}
\usage{
move_example(folder, file)
}
\arguments{
\item{folder}{folder to move template to}

\item{file}{name of new file}
}
\description{
move example
}
