% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_examiners_report.R
\name{parse_examiners_report}
\alias{parse_examiners_report}
\title{parse examiners report}
\usage{
parse_examiners_report(file)
}
\arguments{
\item{file}{examiners report}
}
\value{
tibble of examiners reports
}
\description{
parse examiners report
}
