% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_grade_roster.R
\name{parse_grade_roster}
\alias{parse_grade_roster}
\title{parse_grade_roster}
\usage{
parse_grade_roster(filename, year)
}
\arguments{
\item{filename}{CSV or XLSX grade-roster}

\item{year}{year of offering}
}
\value{
mark-obj
}
\description{
parse_grade_roster
}
