% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gradebook.R
\name{parse_gradebook}
\alias{parse_gradebook}
\title{parse gradebook}
\usage{
parse_gradebook(file, year)
}
\arguments{
\item{file}{gradebook CSV}

\item{year}{year of offering}
}
\value{
mark-object
}
\description{
parse gradebook
}
