% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gradebook_ao.R
\name{parse_gradebook_ao}
\alias{parse_gradebook_ao}
\title{parse gradebook ao}
\usage{
parse_gradebook_ao(file, year)
}
\arguments{
\item{file}{gradebook CSV}

\item{year}{year of offering}
}
\value{
assessment-object
}
\description{
parse gradebook ao
}
