% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gradebook_folder.R
\name{parse_gradebook_folder}
\alias{parse_gradebook_folder}
\title{parse gradebook folder}
\usage{
parse_gradebook_folder(folder, year, type = "mark-obj")
}
\arguments{
\item{folder}{folder of gradebooks}

\item{year}{year of gradebooks}

\item{type}{type of object assessment or mark}
}
\value{
mark-obj or assessment-obj
}
\description{
parse gradebook folder
}
