% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_graderoster.R
\name{parse_graderoster}
\alias{parse_graderoster}
\title{parse_graderoster}
\usage{
parse_graderoster(filename, year)
}
\arguments{
\item{filename}{CSV or XLSX grade-roster}

\item{year}{year of offering}
}
\value{
mark-obj
}
\description{
parse_graderoster
}
