% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_graderoster_folder.R
\name{parse_graderoster_folder}
\alias{parse_graderoster_folder}
\title{parse graderoster folder}
\usage{
parse_graderoster_folder(folder, year)
}
\arguments{
\item{folder}{folder of grade-rosters}

\item{year}{year of grade-rosters}
}
\value{
mark-obj
}
\description{
parse graderoster folder
}
