% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_peoplesoft.R
\name{parse_peoplesoft}
\alias{parse_peoplesoft}
\title{parse_peoplesoft}
\usage{
parse_peoplesoft(file)
}
\arguments{
\item{file}{a peoplesoft file}
}
\value{
mark tibble
}
\description{
parse_peoplesoft
}
