% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_peoplesoft_folder.R
\name{parse_peoplesoft_folder}
\alias{parse_peoplesoft_folder}
\title{parse peoplesoft folder}
\usage{
parse_peoplesoft_folder(folder)
}
\arguments{
\item{folder}{folder of peoplesoft files}
}
\value{
mark-obj
}
\description{
parse peoplesoft folder
}
