% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_xlsx.R
\name{parse_xlsx}
\alias{parse_xlsx}
\title{parse_xlsx}
\usage{
parse_xlsx(filename, year, term)
}
\arguments{
\item{filename}{xlsx spreadsheet}

\item{year}{year of offering}

\item{term}{term of offering}
}
\value{
mark-obj
}
\description{
parse_xlsx
}
