% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_xlsx_folder.R
\name{parse_xlsx_folder}
\alias{parse_xlsx_folder}
\title{parse xlsx folder}
\usage{
parse_xlsx_folder(folder, year, term)
}
\arguments{
\item{folder}{folder of xlsx}

\item{year}{year of xlsx}

\item{term}{term of offering}
}
\value{
mark-obj or assessment-obj
}
\description{
parse xlsx folder
}
