% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assessment_bp.R
\name{plot_assessment_bp}
\alias{plot_assessment_bp}
\title{plot assessment box-plot}
\usage{
plot_assessment_bp(assessment_obj, assessment_list)
}
\arguments{
\item{assessment_obj}{assessment object}

\item{assessment_list}{list of assessments to display}
}
\value{
plot of marks
}
\description{
plot assessment box-plot
}
\examples{
assessment_obj <- assessment_obj |> filter_assessment("DISCP-1234", 2024, "Sem 1")
AL <- assessment_obj |> get_assessment_list() 
plot_assessment_bp(assessment_obj, AL)
}
