% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff.R
\name{plot_diff}
\alias{plot_diff}
\title{plot difference}
\usage{
plot_diff(augmented_mark_obj, course_id, year, term)
}
\arguments{
\item{augmented_mark_obj}{augmented mark object}

\item{course_id}{course ID}

\item{year}{year of offering}

\item{term}{term of offering}
}
\value{
ggplot plot
}
\description{
plot difference
}
