% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fail_rate.R
\name{plot_fail_rate}
\alias{plot_fail_rate}
\title{plot_fail_rate}
\usage{
plot_fail_rate(mark_obj, course_id, year, term, MS, MSPR, ASPR)
}
\arguments{
\item{mark_obj}{standard mark tibble}

\item{course_id}{course-id}

\item{year}{year of latest results}

\item{term}{term of latest results}

\item{MS}{number of medical supps}

\item{MSPR}{medical supp pass rate}

\item{ASPR}{academic supp pass rate}
}
\value{
line-plot
}
\description{
Plots a line plot of the fail rate
}
