% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grade_bc.R
\name{plot_grade_bc}
\alias{plot_grade_bc}
\title{plot_grade_bc}
\usage{
plot_grade_bc(mark_obj, course_id, term, all_grades = FALSE)
}
\arguments{
\item{mark_obj}{standard mark tibble}

\item{course_id}{course-id}

\item{term}{term}

\item{all_grades}{Show all grades}
}
\value{
bar-chart
}
\description{
Plots a bar-chart of grades for mark-obj
}
\examples{
plot_grade_bc(mark_obj, course_id = "DISCP-1001", term = "Sem 1")
plot_grade_bc(mark_obj, course_id = c("DISCP-1001", "DISCP-2001"), term = "Sem 1")
}
