% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intersection.R
\name{plot_intersection}
\alias{plot_intersection}
\title{plot intersection}
\usage{
plot_intersection(intersection_obj)
}
\arguments{
\item{intersection_obj}{intersection object}
}
\value{
plots of pairwise scatter
}
\description{
plot intersection
}
