% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marks.R
\name{plot_marks}
\alias{plot_marks}
\title{plot marks}
\usage{
plot_marks(mark_obj, course_id, year, term)
}
\arguments{
\item{mark_obj}{mark object}

\item{course_id}{course ID}

\item{year}{year of offering}

\item{term}{term of offering}
}
\value{
ggplot plot
}
\description{
plot marks
}
