% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_fail_rate.R
\name{predict_fail_rate}
\alias{predict_fail_rate}
\title{predict_fail_rate}
\usage{
predict_fail_rate(NF, MS, AS, MSPR = 0.8, ASPR = 0.5)
}
\arguments{
\item{NF}{number of fails based on grades}

\item{MS}{number of medical supps}

\item{AS}{number of academic supps}

\item{MSPR}{medical supp expected pass rate}

\item{ASPR}{academic supp expected pass rate}
}
\value{
Predicted number of fails post supps.
}
\description{
predict_fail_rate
}
\examples{
predict_fail_rate(10, 2, 2)
}
