% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_assessment.R
\name{sim_assessment}
\alias{sim_assessment}
\title{sim assessment}
\usage{
sim_assessment(students)
}
\arguments{
\item{students}{students object}
}
\value{
assessment object
}
\description{
sim assessment
}
\examples{
sim_students(2) |> sim_assessment()
}
