% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_degree.R
\name{sim_marks}
\alias{sim_marks}
\title{Simulate marks}
\usage{
sim_marks(students, start_year)
}
\arguments{
\item{students}{students object}

\item{start_year}{year started degree}
}
\value{
mark object
}
\description{
Simulate marks
}
\examples{
sim_students(2) |> sim_marks(2000)
}
