% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_student.R
\name{sim_student}
\alias{sim_student}
\title{sim_student}
\usage{
sim_student(id_length = 7)
}
\arguments{
\item{id_length}{length of ID}
}
\value{
student object
}
\description{
sim_student
}
\examples{
sim_student(id_length = 5)
}
