% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_student.R
\name{sim_students}
\alias{sim_students}
\title{sim students}
\usage{
sim_students(n_students = 10, ...)
}
\arguments{
\item{n_students}{number of students to simulate}

\item{...}{extra variables}
}
\value{
students object
}
\description{
sim students
}
\examples{
sim_students(id_length = 3)
}
