% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_course_names.R
\name{tab_course_names}
\alias{tab_course_names}
\title{tab course names}
\usage{
tab_course_names(mark_obj)
}
\arguments{
\item{mark_obj}{mark-object}
}
\value{
list of course_ids and latest name
}
\description{
tab course names
}
