% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{term_codes}
\alias{term_codes}
\title{term_codes}
\format{
A data frame with 300 rows and 2 variables:
\describe{
\item{term}{Term code}
\item{name}{Term name}
}
}
\source{
"University of Adelaide"
}
\usage{
term_codes
}
\description{
A dataset containing term codes for peoplesoft
}
\keyword{datasets}
