% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukelatorApp.R
\name{tukelatorApp}
\alias{tukelatorApp}
\title{tukelator app}
\usage{
tukelatorApp(mark_obj, assessment_obj, examiners_report, term = "Sem 1")
}
\arguments{
\item{mark_obj}{mark object tibble}

\item{assessment_obj}{assessment object tibble}

\item{examiners_report}{examiners report object}

\item{term}{selected term}
}
\value{
shiny app
}
\description{
tukelator app
}
\examples{
\dontrun{
tukelatorApp(example_marks)
}
}
